add CORS headers
CORS isn't a security precaution to prevent requests from being processed by a server. it
is instead a mechanism to instruct browsers that they can relax their usual cross-origin
security policies a bit. it just so happens that this prevents some requests from being
processed if the browser policy sends a preflight first and does not get permission to
continue.

i'm being permissive with the paths that respond to CORS preflights, because gorilla
gives me no choice and CORS isn't for server-side security anyway.

i'm being permissive with the origin's protocol, because APP_DOMAINS is not specific and
that's what rack-cors does in the ruby server anyway.

i'm being conservative with allowed headers, because there's no "any" option and i'm
unsure what headers will matter.