KeyStoreRotater optimistically generates a key
for a single server environment, this should be straight-up faster
without the extra coordination costs of the WLock.

in a multiple server environment, this trades coordination costs (e.g.
lock waits) and failure modes (lock wait timeouts) for a reliable cpu
cost every hour (with default config) on every node.