Introduce wrapper type for rsa.PrivateKey with JWK
In various places we compute the JWK thumbprint and in get_jwks.go we
use the entire key. We compute these as needed - possibly multiple
times. This change introduces a wrapper type private.Key that wraps up
the rsa.PrivateKey. It allows:

- Single computation of JWK thumbprint and key
- Simplifies logic in get_jwk.go
- Centralises logic for generating RSA key
- Makes it easy to add a KeyStore backed JWKProvider (useful for using
authn-go's verifier when using authn-server as a library)
- Makes JWK available from KeyStore (when using as library)
- Provides a potential place to abstract over different key types (e.g.
ed25519)

Signed-off-by: Silas Davis <silas@monax.io>