dep: bump go-redis (v6.15.2 => v8.11.4)
This patch bumps the go-redis library to v8, which resolves an issue
when authn-server is configured with a non-default Redis user.

v8 notably introduces the use of the stdlib's context package to its
client functions. This patch sidesteps a decision on whether the blob
and refresh token stores should also be context-aware by using
`context.TODO()` as needed.

Closes https://github.com/keratin/authn-server/issues/186.